requirements.txt
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
bitarray==2.6.0
bitstring==4.0.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
crc32c==2.3.post0
Deprecated==1.2.14
distlib==0.3.8
filelock==3.13.1
frozenlist==1.4.1
identify==2.5.34
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
multidict==6.0.5
nodeenv==1.8.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyfiglet==1.0.2
PyNaCl==1.5.0
pytest==8.0.1
python-dotenv==1.0.1
pytoncenter==0.0.11
PyYAML==6.0.1
six==1.16.0
ticton==0.1.11
tonpy==0.0.0.1.2b0
tonsdk==1.0.13
treelib==1.7.0
tvm-valuetypes==0.0.12
typing_extensions==4.9.0
virtualenv==20.25.0
wrapt==1.16.0
yarl==1.9.4
